+++ src/Routers/InstallationsRouter.js
@@ -50,14 +50,12 @@ export class InstallationsRouter extends ClassesRouter {
     return super.handleDelete(req);
   }
 
-  getExpressRouter() {
-    let router = new PromiseRouter();
-    router.route('GET','/installations', req => { return this.handleFind(req); });
-    router.route('GET','/installations/:objectId', req => { return this.handleGet(req); });
-    router.route('POST','/installations', req => { return this.handleCreate(req); });
-    router.route('PUT','/installations/:objectId', req => { return this.handleUpdate(req); });
-    router.route('DELETE','/installations/:objectId', req => { return this.handleDelete(req); });
-    return router;
+  mountRoutes() {
+    this.route('GET','/installations', req => { return this.handleFind(req); });
+    this.route('GET','/installations/:objectId', req => { return this.handleGet(req); });
+    this.route('POST','/installations', req => { return this.handleCreate(req); });
+    this.route('PUT','/installations/:objectId', req => { return this.handleUpdate(req); });
+    this.route('DELETE','/installations/:objectId', req => { return this.handleDelete(req); });
   }
 }